feat: make ball and paddles a little more random feat: reduce max velocity of paddles to prevent model from avoiding the paddles docs: couple other thoughts on features
