+++ model.py
@@ -17,7 +17,7 @@ class PongDataset(IterableDataset):
 
         self.engine_config = EngineConfig()
         # increase max velocity of paddle to even out misses vs hits
-        self.engine_config.set_paddle_factory(RandomPaddleFactory(max_velocity=0.2))
+        self.engine_config.set_paddle_factory(RandomPaddleFactory(max_velocity=0.009))
         self.generator = state_generator
         self.count = count
 
@@ -36,10 +36,10 @@ class PongDataset(IterableDataset):
         self.prepare_worker()
         window_size = input_sequence_length + 1
         window = deque(maxlen=window_size)
-        for ball_data, paddle_data, collision_data, score_data in self.generator(self.count,
-                                                                                 engine_config=self.engine_config):
+        for ball_data, paddle_data, collision_data, score_data in self.generator(engine_config=self.engine_config, num_steps=self.count):
             window.append(ball_data + paddle_data + collision_data + score_data)
             if len(window) == window_size:
+                # print(list(window))
                 states = np.array(window)
                 next_state = np.array(ball_data + collision_data + score_data)
                 yield states[:input_sequence_length], next_state
@@ -69,12 +69,17 @@ class BasePongModel(nn.Module, ABC):
         # Linear layer to expand input from 10 to 64 dimensions
         self.fc_feature_expansion = nn.Linear(input_size, hidden_size)
 
+        # self.intermediate_head = nn.Sequential(
+        #     nn.ReLU(),
+        #     nn.Linear(hidden_size, hidden_size * 4)
+        # )
+
         self.regression_head = nn.Sequential(
             nn.Tanh(),
             nn.Linear(hidden_size, output_size)
         )
         self.classification_head = nn.Sequential(
-            nn.ReLU(),
+            nn.Tanh(),
             nn.Linear(hidden_size, discrete_output_size),
         )
 
@@ -87,6 +92,7 @@ class BasePongModel(nn.Module, ABC):
         # out, _ = self.lstm(x)
         # Use the last output of the sequence for prediction
         x = self._forward(x)
+        # x = self.intermediate_head(x)
         regression_states = self.regression_head(x)  # / temperature
         classification_states = self.classification_head(x) / discrete_temperature
         return regression_states, classification_states
