+++ model_configuration.py
@@ -3,7 +3,7 @@ device = "cuda"
 
 # model parameters
 input_size = 16
-hidden_size = 128
+hidden_size = 64
 output_size = 4
 discrete_output_size = 6
 num_layers = 2
